add support for SIGUSR1 to trigger flush
